Fix tarball dependency extraction implementation
When installing a tarball dependency, Qi would intentionally name the
downloaded tarball: <package>-<version>.tar.gz.

When Qi extracted it, the archive's header might contain a different
name, like: <package>-0.9.4.tar.gz.  So, Qi was not able to find the
package and would error out.

Now, if the extracted archive is not in the expected location, we move
it to there so 1) Qi can load it, and 2) the canonical name/version
pathnames are maintained.