packages: error on broken dependencies
Instead of skipping broken dependencies and adding them
to *qi-broken-dependencies* or *qi-broken-trans-dependencies*, just
error out.

A broken dependency more likely than not means something isn't going to
work until it's fixed; also SBCL is very verbose and it's difficult to
see why a dependency is broken if we continue compiling all the other
packages.